+++ public/app/plugins/datasource/influxdb/components/editor/annotation/AnnotationEditor.tsx
@@ -1,11 +1,10 @@
 import React, { useState } from 'react';
 
-import { QueryEditorProps } from '@grafana/data';
-import { InlineFormLabel, Input } from '@grafana/ui';
+import { QueryEditorProps } from '@grafana/data/src';
+import { InlineFormLabel, Input } from '@grafana/ui/src';
 
-import { InfluxQuery, InfluxOptions } from '../types';
-
-import InfluxDatasource from './../datasource';
+import InfluxDatasource from '../../../datasource';
+import { InfluxOptions, InfluxQuery } from '../../../types';
 
 export const AnnotationEditor = (props: QueryEditorProps<InfluxDatasource, InfluxQuery, InfluxOptions>) => {
   const { query, onChange } = props;
