+++ public/app/plugins/datasource/influxdb/components/editor/query/QueryEditor.tsx
@@ -1,20 +1,20 @@
 import { css } from '@emotion/css';
 import React from 'react';
 
-import { QueryEditorProps } from '@grafana/data';
+import { QueryEditorProps } from '@grafana/data/src';
 
-import InfluxDatasource from '../datasource';
-import { buildRawQuery } from '../queryUtils';
-import { InfluxOptions, InfluxQuery } from '../types';
+import InfluxDatasource from '../../../datasource';
+import { buildRawQuery } from '../../../queryUtils';
+import { InfluxOptions, InfluxQuery } from '../../../types';
 
-import { FluxQueryEditor } from './FluxQueryEditor';
-import { QueryEditorModeSwitcher } from './QueryEditorModeSwitcher';
-import { RawInfluxQLEditor } from './RawInfluxQLEditor';
-import { Editor as VisualInfluxQLEditor } from './VisualInfluxQLEditor/Editor';
+import { FluxQueryEditor } from './flux/FluxQueryEditor';
+import { QueryEditorModeSwitcher } from './influxql/QueryEditorModeSwitcher';
+import { RawInfluxQLEditor } from './influxql/code/RawInfluxQLEditor';
+import { VisualInfluxQLEditor as VisualInfluxQLEditor } from './influxql/visual/VisualInfluxQLEditor';
 
 type Props = QueryEditorProps<InfluxDatasource, InfluxQuery, InfluxOptions>;
 
-export const QueryEditor = ({ query, onChange, onRunQuery, datasource, range, data }: Props): JSX.Element => {
+export const QueryEditor = ({ query, onChange, onRunQuery, datasource }: Props) => {
   if (datasource.isFlux) {
     return (
       <div className="gf-form-query-content">
