+++ public/app/plugins/datasource/influxdb/components/editor/query/influxql/InfluxCheatSheet.tsx
