+++ public/app/plugins/datasource/influxdb/components/editor/query/influxql/code/RawInfluxQLEditor.tsx
@@ -1,11 +1,10 @@
 import React, { useId } from 'react';
 
-import { HorizontalGroup, InlineFormLabel, Input, Select, TextArea } from '@grafana/ui';
+import { HorizontalGroup, InlineFormLabel, Input, Select, TextArea } from '@grafana/ui/src';
 
-import { InfluxQuery } from '../types';
-
-import { DEFAULT_RESULT_FORMAT, RESULT_FORMATS } from './constants';
-import { useShadowedState } from './useShadowedState';
+import { InfluxQuery } from '../../../../../types';
+import { DEFAULT_RESULT_FORMAT, RESULT_FORMATS } from '../../../constants';
+import { useShadowedState } from '../../hooks/useShadowedState';
 
 type Props = {
   query: InfluxQuery;
