+++ public/app/plugins/datasource/influxdb/components/editor/query/influxql/utils/toSelectableValue.ts
@@ -1,4 +1,4 @@
-import { SelectableValue } from '@grafana/data';
+import { SelectableValue } from '@grafana/data/src';
 
 export function toSelectableValue<T extends string>(t: T): SelectableValue<T> {
   return { label: t, value: t };
