+++ public/app/plugins/datasource/influxdb/components/editor/query/influxql/visual/FormatAsSection.tsx
@@ -1,13 +1,13 @@
 import { cx } from '@emotion/css';
 import React from 'react';
 
-import { Select } from '@grafana/ui';
+import { Select } from '@grafana/ui/src';
 
-import { ResultFormat } from '../../types';
-import { RESULT_FORMATS } from '../constants';
+import { ResultFormat } from '../../../../../types';
+import { RESULT_FORMATS } from '../../../constants';
+import { unwrap } from '../utils/unwrap';
 
 import { paddingRightClass } from './styles';
-import { unwrap } from './unwrap';
 
 type Props = {
   inputId?: string;
