+++ public/app/plugins/datasource/influxdb/components/editor/query/influxql/visual/OrderByTimeSection.tsx
@@ -1,11 +1,12 @@
 import { cx } from '@emotion/css';
 import React from 'react';
 
-import { SelectableValue } from '@grafana/data';
-import { Select } from '@grafana/ui';
+import { SelectableValue } from '@grafana/data/src';
+import { Select } from '@grafana/ui/src';
+
+import { unwrap } from '../utils/unwrap';
 
 import { paddingRightClass } from './styles';
-import { unwrap } from './unwrap';
 
 type Mode = 'ASC' | 'DESC';
