+++ public/app/plugins/datasource/influxdb/components/editor/query/influxql/visual/VisualInfluxQLEditor.tags.test.tsx
@@ -2,11 +2,11 @@ import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 
-import InfluxDatasource from '../../datasource';
-import * as mockedMeta from '../../influxql_metadata_query';
-import { InfluxQuery } from '../../types';
+import InfluxDatasource from '../../../../../datasource';
+import * as mockedMeta from '../../../../../influxql_metadata_query';
+import { InfluxQuery } from '../../../../../types';
 
-import { Editor } from './Editor';
+import { VisualInfluxQLEditor } from './VisualInfluxQLEditor';
 
 jest.mock('../../influxql_metadata_query', () => {
   return {
@@ -117,14 +117,14 @@ const query: InfluxQuery = {
   measurement: 'cpudata',
 };
 
-describe('InfluxDB InfluxQL Visual Editor field-filtering', () => {
+describe('InfluxDB InfluxQL Visual VisualInfluxQLEditor field-filtering', () => {
   it('should not send fields in tag-structures to metadata queries', async () => {
     const onChange = jest.fn();
     const onRunQuery = jest.fn();
     const datasource: InfluxDatasource = {
       metricFindQuery: () => Promise.resolve([]),
     } as unknown as InfluxDatasource;
-    render(<Editor query={query} datasource={datasource} onChange={onChange} onRunQuery={onRunQuery} />);
+    render(<VisualInfluxQLEditor query={query} datasource={datasource} onChange={onChange} onRunQuery={onRunQuery} />);
 
     await waitFor(() => {});
