+++ public/app/plugins/datasource/influxdb/components/editor/query/influxql/visual/VisualInfluxQLEditor.test.tsx
@@ -1,11 +1,11 @@
 import { render, waitFor } from '@testing-library/react';
 import React from 'react';
 
-import InfluxDatasource from '../../datasource';
+import InfluxDatasource from '../../../../../datasource';
+import { InfluxQuery } from '../../../../../types';
 import { getMockDS, getMockDSInstanceSettings } from '../../specs/mocks';
-import { InfluxQuery } from '../../types';
 
-import { Editor } from './Editor';
+import { VisualInfluxQLEditor } from './VisualInfluxQLEditor';
 
 // we mock the @grafana/ui components we use to make sure they just show their "value".
 // we mostly need this for `Input`, because that one is not visible with `.textContent`,
@@ -42,7 +42,7 @@ async function assertEditor(query: InfluxQuery, textContent: string) {
   const datasource: InfluxDatasource = getMockDS(getMockDSInstanceSettings());
   datasource.metricFindQuery = () => Promise.resolve([]);
   const { container } = render(
-    <Editor query={query} datasource={datasource} onChange={onChange} onRunQuery={onRunQuery} />
+    <VisualInfluxQLEditor query={query} datasource={datasource} onChange={onChange} onRunQuery={onRunQuery} />
   );
   await waitFor(() => {
     expect(container.textContent).toBe(textContent);
