+++ public/app/plugins/datasource/influxdb/components/editor/query/influxql/visual/styles.ts
