+++ public/app/plugins/datasource/influxdb/datasource.mock.ts
@@ -12,6 +12,7 @@ type FieldReturnValue = { text: string };
  */
 export class InfluxDatasourceMock {
   constructor(private measurements: Measurements) {}
+
   metricFindQuery(query: string) {
     if (isMeasurementsQuery(query)) {
       return this.getMeasurements();
