+++ public/app/plugins/datasource/influxdb/module.ts
@@ -1,9 +1,9 @@
 import { DataSourcePlugin } from '@grafana/data';
 
-import ConfigEditor from './components/ConfigEditor';
-import InfluxStartPage from './components/InfluxStartPage';
-import { QueryEditor } from './components/QueryEditor';
-import VariableQueryEditor from './components/VariableQueryEditor';
+import ConfigEditor from './components/editor/config/ConfigEditor';
+import { QueryEditor } from './components/editor/query/QueryEditor';
+import InfluxStartPage from './components/editor/query/influxql/InfluxStartPage';
+import VariableQueryEditor from './components/editor/variable/VariableQueryEditor';
 import InfluxDatasource from './datasource';
 
 export const plugin = new DataSourcePlugin(InfluxDatasource)
